chore: rename project from Comfy-xPixArt to Comfy-TinyBreaker
This commit changes the project name from `Comfy-xPixArt` to `Comfy-TinyBreaker`.
The change reflects a significant shift in the project's focus, now
concentrating on the development and experimentation of the TinyBreaker
model.
As part of this transition, nodes dedicated exclusively to PixArt are being
removed, while internal code related to PixArt is still used for supporting
TinyBreaker.
The project no longer aims to implement PixArt-related custom nodes, as ComfyUI
already provides native support for that.