chore: Add example grid and thumbnail images
This commit adds two new image files to the project. The `tinybreaker_grid.jpg`
file contains a grid of example images generated using the TinyBreaker model,
which will be displayed on the main page to showcase its capabilities.
Additionally, the `thumbnail.jpg` file is included as an icon representation
for TinyBreaker.