fix: Potentially resolve #4 by enforcing positional encoding placement in TimestepEmbedder
This commit introduces a potential fix for issue #4, which was displaying
the following error:
"
RuntimeError: Expected all tensors to be on the same device, but found
at least two devices, cpu and cuda:0! (when checking argument for argument
mat1 in method wrapper_CUDA_addmm)
"
The change forces the positional encoding to reside on the same device
as the MLP within the `TimestepEmbedder` class of `pixart_model.py`.