refactor: Enable injection of custom torch.nn modules in PixArt model
This commit refactors the custom implementation of PixArt within the
project to facilitate easy integration with a customized or optimized
version of `torch.nn`.

The primary goal is to seamlessly incorporate the specialized `torch.nn`
handling employed by ComfyUI, aiming to leverage its memory management
capabilities for VRAM/RAM optimization and improve integration of
'ComfyUI-TinyBreaker' within the ComfyUI system.