feat: Deprecate "LoadTinyBreakerCheckpoint" and introduce "LoadTinyBreakerCheckpointV2".
Deprecated the "LoadTinyBreakerCheckpoint" node and introduced a new
version, "LoadTinyBreakerCheckpointV2". While the core functionality
remains largely unchanged for now, the V2 version lays the groundwork
for future enhancements, specifically enabling the output of an
embedded CLIP model.

Due to connector changes required for this potential future functionality,
a new version has been created to maintain compatibility with existing
workflows that utilize the node.