chore: Move example workflow and prepare for v0.3.0
This commit relocates the `tinybreaker_style_creator.json` workflow
from the `tools` directory to the root of the `workflows` directory.
This change anticipates the workflow's inclusion as a primary workflow
in the next release (v0.3.0).

To reflect this upcoming version, the filename has been updated
to `v03_tinybreaker_style_creator.json`.