fix: Correct sigma calculation function
This commit addresses an error in the internal function responsible
for calculating sigma values during sampling. The issue prevented
the function from considering the `steps_end` parameter, which
specifies the end of the sigma range in the sampling.

Fixing this error resulted in adjustments to style parameters
(particularly `steps` and `steps_end`) to reproduce images
generated when the error was present

 Additionally, the use of `discard_last_sigma` during upscaling has
been eliminated, as it was initially implemented due to an issue that
no longer exists with the corrected function.